Move generate and send incremental export to utils
