+++ corehq/apps/export/tasks.py
@@ -12,7 +12,7 @@ from celery.schedules import crontab
 from celery.task import periodic_task, task
 from corehq.apps.data_dictionary.util import add_properties_to_data_dictionary
 from corehq.apps.export.exceptions import RejectedStaleExport
-from corehq.apps.export.utils import get_export
+from corehq.apps.export.utils import get_export, generate_and_send_incremental_export
 from corehq.apps.users.models import CouchUser
 from corehq.blobs import CODES, get_blob_db
 from corehq.celery_monitoring.signals import get_task_time_to_start
@@ -28,14 +28,9 @@ from .dbaccessors import (
     get_properly_wrapped_export_instance,
 )
 from .export import (
-    ExportFile,
-    _get_export_query,
     get_export_file,
-    get_export_writer,
     rebuild_export,
-    write_export_instance,
 )
-from .filters import ServerModifiedOnRangeFilter
 from .models.incremental import IncrementalExport
 from .models.new import EmailExportWhenDoneRequest
 from .system_properties import MAIN_CASE_TABLE_PROPERTIES
@@ -231,72 +226,4 @@ def process_incremental_export(incremental_export_id):
     incremental_export = IncrementalExport.objects.get(id=incremental_export_id)
     last_valid_checkpoint = incremental_export.last_valid_checkpoint
     last_doc_date = last_valid_checkpoint.last_doc_date if last_valid_checkpoint else None
-
-    checkpoint = _generate_incremental_export(incremental_export, last_doc_date)
-    if checkpoint:
-        _send_incremental_export(incremental_export, checkpoint)
-
-
-def _generate_incremental_export(incremental_export, last_doc_date=None):
-    export_instance = incremental_export.export_instance
-    export_instance.export_format = Format.UNZIPPED_CSV  # force to unzipped CSV
-
-    # Remove the date period from the ExportInstance, since this is added automatically by Daily Saved exports
-    export_instance.filters.date_period = None
-    filters = export_instance.get_filters()
-    if last_doc_date:
-        filters.append(ServerModifiedOnRangeFilter(gt=last_doc_date))
-
-    class LastDocTracker:
-        def __init__(self, doc_iterator):
-            self.doc_iterator = doc_iterator
-            self.last_doc = None
-            self.doc_count = 0
-
-        def __iter__(self):
-            for doc in self.doc_iterator:
-                self.last_doc = doc
-                self.doc_count += 1
-                yield doc
-
-    with TransientTempfile() as temp_path, metrics_track_errors('generate_incremental_exports'):
-        writer = get_export_writer([export_instance], temp_path, allow_pagination=False)
-        with writer.open([export_instance]):
-            query = _get_export_query(export_instance, filters)
-            query = query.sort('server_modified_on')  # reset sort to this instead of opened_on
-            docs = LastDocTracker(query.run().hits)
-            write_export_instance(writer, export_instance, docs)
-
-        export_file = ExportFile(writer.path, writer.format)
-
-        if docs.doc_count <= 0:
-            return
-
-        new_checkpoint = incremental_export.checkpoint(
-            docs.doc_count, docs.last_doc.get('server_modified_on')
-        )
-
-        with export_file as file_:
-            db = get_blob_db()
-            db.put(
-                file_,
-                domain=incremental_export.domain,
-                parent_id=new_checkpoint.blob_parent_id,
-                type_code=CODES.data_export,
-                key=str(new_checkpoint.blob_key),
-                timeout=24 * 60
-            )
-    return new_checkpoint
-
-
-def _send_incremental_export(export, checkpoint):
-    requests = _get_requests(checkpoint, export)
-    headers = {
-        'Accept': 'application/json'
-    }
-    files = {'file': (checkpoint.filename, checkpoint.get_blob(), 'text/csv')}
-    requests.post(uri='', files=files, headers=headers)
-
-
-def _get_requests(checkpoint, export):
-    return export.connection_settings.get_requests(checkpoint.id, checkpoint.log_request)
+    generate_and_send_incremental_export(incremental_export, last_doc_date)
